+++ packages/react-liveness/src/components/FaceLivenessDetector/service/utils/__tests__/streamProvider.test.ts
@@ -4,7 +4,6 @@ import 'blob-polyfill';
 
 import { TextDecoder } from 'util';
 import { RekognitionStreamingClient } from '@aws-sdk/client-rekognitionstreaming';
-import { Amplify } from 'aws-amplify';
 
 import { LivenessStreamProvider } from '../streamProvider';
 import { VideoRecorder } from '../videoRecorder';
@@ -20,19 +19,24 @@ Object.defineProperty(window, 'TextDecoder', {
   value: TextDecoder,
 });
 
-const mockGet = jest.fn().mockImplementation(() => {
+jest.mock('aws-amplify/auth', () => {
+  const originalModule = jest.requireActual('aws-amplify/auth');
   return {
-    accessKeyId: 'accessKeyId',
-    sessionToken: 'sessionTokenId',
-    secretAccessKey: 'secretAccessKey',
-    identityId: 'identityId',
-    authenticated: true,
-    expiration: new Date(),
+    ...originalModule,
+    fetchAuthSession: jest.fn().mockImplementation(() => {
+      return Promise.resolve({
+        credentials: {
+          accessKeyId: 'accessKeyId',
+          sessionToken: 'sessionTokenId',
+          secretAccessKey: 'secretAccessKey',
+          identityId: 'identityId',
+          authenticated: true,
+          expiration: new Date(),
+        },
+      });
+    }),
   };
 });
-
-// @todo-migration remove cast and fix mock if needed
-(Amplify as any).Credentials.get = mockGet;
 
 let SWITCH = false;